django_tea_do/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
coreapi==2.3.3